+++ shared/src/business/useCases/caseAssociationRequest/submitCaseAssociationRequestInteractor.js
@@ -1,4 +1,4 @@
-const { Case } = require('../../entities/Case');
+const { Case } = require('../../entities/cases/Case');
 const { UnauthorizedError } = require('../../../errors/errors');
 
 /**
