+++ shared/src/business/useCases/getCaseInteractor.js
@@ -2,7 +2,7 @@ const {
   GET_CASE,
   isAuthorized,
 } = require('../../authorization/authorizationClientService');
-const { Case } = require('../entities/Case');
+const { Case } = require('../entities/cases/Case');
 const { NotFoundError, UnauthorizedError } = require('../../errors/errors');
 
 /**
