+++ shared/src/business/useCases/getCasesByUserInteractor.js
@@ -1,4 +1,4 @@
-const { Case } = require('../entities/Case');
+const { Case } = require('../entities/cases/Case');
 
 /**
  * getCasesByUser
