+++ shared/src/business/useCases/getProcedureTypesInteractor.js
@@ -1,4 +1,4 @@
-const { Case } = require('../entities/Case');
+const { Case } = require('../entities/cases/Case');
 
 /**
  * getProcedureTypes
