+++ shared/src/business/useCases/recallPetitionFromIRSHoldingQueueInteractor.js
@@ -10,7 +10,7 @@ const {
   isAuthorized,
   UPDATE_CASE,
 } = require('../../authorization/authorizationClientService');
-const { Case, STATUS_TYPES } = require('../entities/Case');
+const { Case, STATUS_TYPES } = require('../entities/cases/Case');
 const { Document } = require('../entities/Document');
 
 const {
