+++ web-client/e2e/trialSessionEligibleCasesScenario3.test.js
@@ -5,7 +5,7 @@ import FormData from 'form-data';
 import {
   CASE_CAPTION_POSTFIX,
   STATUS_TYPES,
-} from '../../shared/src/business/entities/Case';
+} from '../../shared/src/business/entities/cases/Case';
 import {
   CATEGORIES,
   CATEGORY_MAP,
