+++ web-client/src/index.dev.js
@@ -1,7 +1,7 @@
 import '@babel/polyfill';
 import Devtools from 'cerebral/devtools';
 
-import { Case } from '../../shared/src/business/entities/Case';
+import { Case } from '../../shared/src/business/entities/cases/Case';
 import { User } from '../../shared/src/business/entities/User';
 import { app } from './app';
 import { applicationContext } from './applicationContext';
