+++ web-client/src/presenter/actions/updateCaseAction.test.js
@@ -1,7 +1,7 @@
 import {
   Case,
   STATUS_TYPES,
-} from '../../../../shared/src/business/entities/Case';
+} from '../../../../shared/src/business/entities/cases/Case';
 import { MOCK_CASE } from '../../../../shared/src/test/mockCase';
 import { presenter } from '../presenter';
 import { runAction } from 'cerebral/test';
